refactor: move createNewTestSuiteAction to dedicated route file
