+++ packages/insomnia/src/ui/routes/actions.tsx
@@ -40,24 +40,6 @@ export function safeToUseInsomniaFileNameWithExt(fileName: string) {
 }
 
 // Test Suite
-export const createNewTestSuiteAction: ActionFunction = async ({ request, params }) => {
-  const { organizationId, workspaceId, projectId } = params;
-  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
-  const formData = await request.formData();
-  const name = formData.get('name');
-  invariant(typeof name === 'string', 'Name is required');
-
-  const unitTestSuite = await models.unitTestSuite.create({
-    parentId: workspaceId,
-    name,
-  });
-
-  window.main.trackSegmentEvent({ event: SegmentEvent.testSuiteCreate });
-
-  return redirect(
-    `/organization/${organizationId}/project/${projectId}/workspace/${workspaceId}/test/test-suite/${unitTestSuite._id}`,
-  );
-};
 
 export const deleteTestSuiteAction: ActionFunction = async ({ params }) => {
   const { organizationId, workspaceId, projectId, testSuiteId } = params;
